move private Points field in sub-classes to Figure abstract class so that instantiated when initializing 1. some problems happened. initializing exception part in sub-class's constructor was moved to factory construct method 'of' and this is right changing? 2. cause each area calculated method refer to Points instance which is in abstract figure class, should be moved to abstract method from sub-classes.
