+++ src/main/java/domain/Line.java
@@ -6,27 +6,17 @@ import static domain.FigureUtil.isLine;
 
 public class Line extends Figure {
 
-    private Points points;
-
     private double area;
 
     private Line(List<String> coordinates) {
-        if (!isLine(coordinates)) {
-            throw new RuntimeException();
-        }
-        points = Points.initCoordinates(coordinates);
+        super(Points.initCoordinates(coordinates));
     }
 
     public static Line of(List<String> coordinates) {
+        if (!isLine(coordinates)) {
+            throw new RuntimeException();
+        }
         return new Line(coordinates);
-    }
-
-    double calculateLine() {
-        return points.calculateLineBwPoints(0, 1);
-    }
-
-    public boolean hasCoordinateAt(int x, int y) {
-        return points.contains(x, y);
     }
 
     public double getArea(){
