+++ src/test/java/domain/LineTest.java
@@ -10,10 +10,5 @@ import static org.junit.Assert.assertEquals;
 
 public class LineTest {
 
-    @Test
-    public void calculateLine(){
-        List<String> coordinates = Arrays.asList("1,1","2,2");
-        Line line = Line.of(coordinates);
-        assertEquals(1.414, line.calculateLine(), 0.001);
-    }
+
 }
